Accounts and permissions
This documentation includes steps to install AD Connect:
In the pop-up dialog, you can either (i) provide an Enterprise Admin credential and let Azure AD Connect create the AD DS account for you, or (ii) create the AD DS account yourself and provide its credential to Azure AD Connect. Once you have selected an option and provide the necessary credentials, click OK to close the pop-up dialog.

Customers are following this document and may not be aware of updates made to > https://docs.microsoft.com/en-us/azure/active-directory/hybrid/reference-connect-accounts-permissions

As of build 1.4.###.# it is no longer supported to use an enterprise admin or a domain admin account as the AD DS Connector account. If you attempt to enter an account that is an enterprise admin or domain admin when specifying use existing account, you will receive an error.